==> 19_rna_calculate_fpkm.ipynb <==
import pandas as pd

rna_featurecounts_df = pd.read_csv('./7_featurecounts_rna/gene_exon_counts.txt', sep='\t', comment='#')

rna_featurecounts_df

rna_featurecounts_df["Geneid"].nunique()

# 提取counts数据和基因长度
gene_lengths = rna_featurecounts_df['Length']
counts = rna_featurecounts_df.iloc[:, 6:]  # 从第七列开始是样本的counts数据

# 筛选包含 "input.bam" 的列
input_counts = counts.loc[:, counts.columns.str.contains('input.bam')]

# 计算每个样本的总mapped reads
total_counts = input_counts.sum(axis=0)

# 计算FPKM
fpkm = input_counts.div(gene_lengths, axis=0)
fpkm = fpkm.div(total_counts / 1e6, axis=1)

# 将 FPKM 数据框的列名加上前缀 "FPKM_"
fpkm.columns = ['FPKM_' + col.split('/')[-1] for col in fpkm.columns]

fpkm

# 合并 FPKM 数据到原始 DataFrame
rna_featurecounts_df = pd.concat([rna_featurecounts_df, fpkm], axis=1)

#### 读取hg38 gene dict

df_hg38 = pd.read_csv("./hg38_111_annotate.txt", sep=' ', header=None, names=["id", "name", "type"])
# filtered_df_hg38 = df_hg38[df_hg38["type"] == "protein_coding"]
filtered_df_hg38 = df_hg38

filtered_df_hg38

rna_featurecounts_df

# 使用merge函数根据Geneid和id进行左连接合并，这样可以保留rna_featurecounts_df中所有行，即使没有匹配的id
merged_df = pd.merge(rna_featurecounts_df, filtered_df_hg38, left_on='Geneid', right_on='id', how='left')

# 选择需要的列，并重命名name列为Gene
result_df = merged_df.rename(columns={'name': 'Gene'})

result_df

result_df = result_df[result_df["type"] == "protein_coding"]

result_df

# 处理'Chr'列，取第一个字符并添加'chr'前缀
result_df['Chr'] = 'chr' + result_df['Chr'].str.split(';').str[0]

# 处理'Strand'列，取第一个字符
result_df['Strand'] = result_df['Strand'].str.split(';').str[0]

result_df = result_df[["Gene", "Length", "Chr", "Strand", "FPKM_BE2-shMETTL14-1-input.bam", "FPKM_BE2-shMETTL14-2-input.bam", "FPKM_BE2-shNC-1-input.bam", "FPKM_BE2-shNC-2-input.bam"]]

# 保存结果到 CSV 文件
output_file = './7_featurecounts_rna/gene_exon_counts_with_fpkm.csv'
result_df.to_csv(output_file, sep='\t', index=False)

print(f"FPKM calculation completed. Output written to {output_file}")
