==> 20_process_rna_m6a_sequence.ipynb <==
import pandas as pd
from tqdm import tqdm

df_be1 = pd.read_csv("./7_featurecounts/BE2-shMETTL14-1-IP_fwd_peaks_peaks_merged.csv")
df_be2 = pd.read_csv("./7_featurecounts/BE2-shMETTL14-1-IP_rev_peaks_peaks_merged.csv")
df_be1["name"] = "shMETTL14-1"
df_be2["name"] = "shMETTL14-1"
df_be1["Strand"] = "-"
df_be2["Strand"] = "+"

df_be3 = pd.read_csv("./7_featurecounts/BE2-shMETTL14-2-IP_fwd_peaks_peaks_merged.csv")
df_be4 = pd.read_csv("./7_featurecounts/BE2-shMETTL14-2-IP_rev_peaks_peaks_merged.csv")
df_be3["name"] = "shMETTL14-2"
df_be4["name"] = "shMETTL14-2"
df_be3["Strand"] = "-"
df_be4["Strand"] = "+"

df_be5 = pd.read_csv("./7_featurecounts/BE2-shNC-1-IP_fwd_peaks_peaks_merged.csv")
df_be6 = pd.read_csv("./7_featurecounts/BE2-shNC-1-IP_rev_peaks_peaks_merged.csv")
df_be5["name"] = "shNC-1"
df_be6["name"] = "shNC-1"
df_be5["Strand"] = "-"
df_be6["Strand"] = "+"

df_be7 = pd.read_csv("./7_featurecounts/BE2-shNC-2-IP_fwd_peaks_peaks_merged.csv")
df_be8 = pd.read_csv("./7_featurecounts/BE2-shNC-2-IP_rev_peaks_peaks_merged.csv")
df_be7["name"] = "shNC-2"
df_be8["name"] = "shNC-2"
df_be7["Strand"] = "-"
df_be8["Strand"] = "+"

df_be_merge = pd.concat([
    df_be1[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]], 
    df_be2[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be3[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be4[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be5[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be6[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be7[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]],
    df_be8[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "kmers", "Site"]]
    ], axis=0, ignore_index=True)

df_be_merge.to_csv("./7_featurecounts/BE2_merged_all.csv", sep='\t', index=False)

gene_exon_fpkm_df = pd.read_csv('./7_featurecounts_rna/gene_exon_counts_with_fpkm.csv', sep='\t')
m6a_merged_cpm_df = pd.read_csv('./7_featurecounts/BE2_merged_all.csv', sep='\t')
mane_gene_feature_df = pd.read_csv("./7_MANE_Gene_mRNA_feature_df.csv", sep=',')

gene_exon_fpkm_df

mane_gene_feature_df

m6a_merged_cpm_df

#### 根据mane gene数据对gene_exon_fpkm数据进行提取合并

# 只选择需要的列进行merge
mane_gene_feature_df_subset = mane_gene_feature_df[["Gene", "Chr", "Strand", "Stitched_mRNA_sequence", "5'UTR", "CDS", "3'UTR"]]

# 按行去重
mane_gene_feature_df_subset = mane_gene_feature_df_subset.drop_duplicates()

# 进行合并操作，保留gene_exon_fpkm_df中的所有行
merge_rna_mane_gene = gene_exon_fpkm_df.merge(mane_gene_feature_df_subset, on=["Gene", "Chr", "Strand"], how="left")

# 删除没有匹配到的行（即包含NaN值的行）
merge_rna_mane_gene = merge_rna_mane_gene.dropna(subset=["Stitched_mRNA_sequence", "5'UTR", "CDS", "3'UTR"])

merge_rna_mane_gene.to_csv("./data/merge_rna_mane_gene.csv")

merge_rna_mane_gene

#### 根据mane gene数据对m6a_merged_cpm_df进行提取合并

# 初始化新的列
m6a_merged_cpm_df["Gene"] = None
m6a_merged_cpm_df["m6a_position"] = None

# 将Site列的字符串转为列表
m6a_merged_cpm_df['Site'] = m6a_merged_cpm_df['Site'].apply(eval)

# 遍历m6a_merged_cpm_df
for idx, row in tqdm(m6a_merged_cpm_df.iterrows()):
    gene_matches = []
    positions = []
    
    match = mane_gene_feature_df[
        (mane_gene_feature_df['Chr'] == row['Chr']) &
        (mane_gene_feature_df['Strand'] == row['Strand'])
    ]

    for site in row['Site']:

        for _, gene_row in match.iterrows():
            gene_start = gene_row['Start']
            gene_end = gene_row['End']

            if gene_start <= site <= gene_end:
                print(f"{site} is match")
                
                gene_matches.append(gene_row['Gene'])
                positions.append(site - gene_start)
                break
         
                
    if gene_matches:
        m6a_merged_cpm_df.at[idx, 'Gene'] = gene_matches[0]
        m6a_merged_cpm_df.at[idx, 'm6a_position'] = positions

m6a_merged_cpm_df

m6a_merged_cpm_df.to_csv("./data/merge_m6a_mane_gene.csv")
