==> 21_read_rna_m6a_sequence.ipynb <==
import pandas as pd


merge_rna_mane_gene_df = pd.read_csv("./data/merge_rna_mane_gene.csv")

merge_rna_mane_gene_df

m6a_merged_cpm_df = pd.read_csv("./data/merge_m6a_mane_gene.csv")

m6a_merged_cpm_df_drop = m6a_merged_cpm_df.dropna()
m6a_merged_cpm_df_drop[["name", "Chr", "Start", "End", "Strand", "Length", "IP/input", "Site", "Gene", "m6a_position"]]

merge_gene_m6a_df = pd.merge(merge_rna_mane_gene_df, m6a_merged_cpm_df_drop, on="Gene", how="left")

merge_gene_m6a_df

merge_gene_m6a_df_drop = merge_gene_m6a_df.dropna()
merge_gene_m6a_df_drop[["Gene", "name", "IP/input", "m6a_position", "FPKM_BE2-shMETTL14-1-input.bam", "FPKM_BE2-shMETTL14-2-input.bam", "FPKM_BE2-shNC-1-input.bam", "FPKM_BE2-shNC-2-input.bam", "Stitched_mRNA_sequence", "5'UTR", "CDS", "3'UTR"]]

# 定义一个函数，根据'name'列的值选择相应的'FPKM'列的值
def get_rna_expression(row):
    # 构建FPKM列名
    fpkm_col_name = f"FPKM_BE2-{row['name']}-input.bam"
    # 返回对应的FPKM值
    return row[fpkm_col_name]

merge_gene_m6a_df_drop['RNA_express'] = merge_gene_m6a_df_drop.apply(get_rna_expression, axis=1)

merge_gene_m6a_df_drop[["Gene", "IP/input", "m6a_position", "RNA_express", "Stitched_mRNA_sequence", "5'UTR", "CDS", "3'UTR"]]

#### 分析下BE2_merge的数据

m6a_merged_cpm_df = pd.read_csv('./7_featurecounts/BE2_merged_all.csv', sep='\t')

m6a_merged_cpm_df

m6a_merged_cpm_df['Site'] = m6a_merged_cpm_df['Site'].apply(eval)

m6a_merged_cpm_df["Site"].apply(lambda x: x != []).sum()

[item - m6a_merged_cpm_df["Start"].iloc[10] for item in m6a_merged_cpm_df["Site"].iloc[10]]

aa = m6a_merged_cpm_df["kmers"].iloc[10]

print(aa[43:43+5], aa[115:115+5], aa[186:186+5], aa[272:272+5])
print(aa[43+2], aa[115+2], aa[186+2], aa[272+2])
